Improved token generation logic for refresh and confirmation tokens + Refactored project architecture for better organization and readability.
