--- Auth/src/main/java/com/example/auth/AuthenticationRequest.java
@@ -1,13 +1,0 @@
-package com.example.auth;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class AuthenticationRequest {
-    private String email;
-    private String password;
-}
