+++ Auth/src/main/java/com/example/dto/request/AuthenticationRequest.java
@@ -1,14 +1,13 @@
-package com.example.auth;
+package com.example.dto.request;
 
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class AuthenticationResponse {
-    private String accessToken;
+public class AuthenticationRequest {
+    private String email;
+    private String password;
 }
