+++ Auth/src/main/java/com/example/dto/response/AuthenticationResponse.java
@@ -1,0 +1,15 @@
+package com.example.dto.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class AuthenticationResponse {
+    private String accessToken;
+    private String refreshToken;
+}
