+++ Auth/src/main/java/com/example/services/AuthenticationService.java
@@ -1,9 +1,10 @@
-package com.example.auth;
+package com.example.services;
 
+import com.example.dto.request.AuthenticationRequest;
+import com.example.dto.response.AuthenticationResponse;
 import com.example.config.JwtService;
 import com.example.dto.DoctorDTO;
 import com.example.dto.PatientDTO;
-import com.example.dto.UserDetailsImpl;
 import com.example.entities.Doctor;
 import com.example.entities.Patient;
 import com.example.entities.Token;
@@ -15,19 +16,13 @@ import com.example.mapper.AutoPatientMapper;
 import com.example.mapper.AutoUserMapper;
 import com.example.repositories.TokenRepository;
 import com.example.repositories.UserRepository;
-import com.example.services.EmailSenderService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.ResponseEntity;
 import org.springframework.mail.MailException;
 import org.springframework.mail.SimpleMailMessage;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.scheduling.annotation.Async;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationServiceException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -43,14 +38,16 @@ public class AuthenticationService {
     @Autowired
     AutoPatientMapper autoPatientMapper;
     @Autowired
+    ConfirmationTokenService confirmationTokenService;
+    @Autowired
+    RefreshTokenService refreshTokenService;
+    @Autowired
     AutoDoctorMapper autoDoctorMapper;
     @Autowired
     AutoUserMapper autoUserMapper;
     private final AuthenticationManager authenticationManager;
     @Autowired
     private final EmailSenderService emailSenderService;
-    @Value("${application.security.jwt.refresh-token.expiration}")
-    private long refreshExpiration;
     private final Logger logger = LoggerFactory.getLogger(AuthenticationService.class);
 
     public AuthenticationResponse register(PatientDTO request) {
@@ -63,10 +60,7 @@ public class AuthenticationService {
             patient.setPassword(passwordEncoder.encode(patient.getPassword()));
             repository.save(patient);
             var jwtToken = jwtService.generateToken(autoUserMapper.toDto(patient));
-//            var refreshToken = jwtService.generateRefreshToken(autoUserMapper.toDto(patient));
-            Token confirmToken = jwtService.generateConfirmToken(autoUserMapper.toDto(patient));
-            tokenRepository.save(confirmToken);
-//            tokenRepository.save(refreshToken);
+            Token confirmToken = confirmationTokenService.generateConfirmToken(autoUserMapper.toDto(patient));
 
             try {
                 SimpleMailMessage mailMessage = new SimpleMailMessage();
@@ -78,9 +72,7 @@ public class AuthenticationService {
 
                 emailSenderService.sendEmail(mailMessage);
             } catch (MailException ex) {
-                // Log the exception
                 ex.printStackTrace();
-                // Handle the exception (e.g., notify the user about the issue)
             }
 
             return AuthenticationResponse.builder()
@@ -99,9 +91,7 @@ public class AuthenticationService {
             doctor.setPassword(passwordEncoder.encode(doctor.getPassword()));
             repository.save(doctor);
             var jwtToken = jwtService.generateToken(autoUserMapper.toDto(doctor));
-            Token confirmToken = jwtService.generateConfirmToken(autoUserMapper.toDto(doctor));
-            tokenRepository.save(confirmToken);
-//            jwtService.generateRefreshToken(autoUserMapper.toDto(doctor));
+            Token confirmToken = confirmationTokenService.generateConfirmToken(autoUserMapper.toDto(doctor));
 
             try {
                 SimpleMailMessage mailMessage = new SimpleMailMessage();
@@ -113,9 +103,7 @@ public class AuthenticationService {
 
                 emailSenderService.sendEmail(mailMessage);
             } catch (MailException ex) {
-                // Log the exception
                 ex.printStackTrace();
-                // Handle the exception (e.g., notify the user about the issue)
             }
 
             return AuthenticationResponse.builder()
@@ -129,6 +117,7 @@ public class AuthenticationService {
         Token confirmToken = tokenRepository.findByUserAndTokenType(user, TokenType.CONFIRM_ACCOUNT);
         if (confirmToken != null){
             logger.error("ERROR: Account not confirmed!");
+            return null;
         }else {
             authenticationManager.authenticate(
                     new UsernamePasswordAuthenticationToken(
@@ -136,10 +125,13 @@ public class AuthenticationService {
                             request.getPassword()
                     )
             );
+            var jwtToken = jwtService.generateToken(autoUserMapper.toDto(user));
+            var refreshToken = refreshTokenService.generateRefreshToken(autoUserMapper.toDto(user));
+
+            return AuthenticationResponse.builder()
+                    .accessToken(jwtToken)
+                    .refreshToken(refreshToken.toString())
+                    .build();
         }
-        var jwtToken = jwtService.generateToken(autoUserMapper.toDto(user));
-        return AuthenticationResponse.builder()
-                .accessToken(jwtToken)
-                .build();
     }
 }
